add authentication and fix get device data

+++ server/model/home.py
@@ -12,15 +12,22 @@ class HomeModel():
             query["_id"] = home_id
         try:
             home_list = self._database.fetch_data(config.database.DOC_HOME_LIST, query)
-            if len(home_list) < 1:
-                raise RecordFindError(home_id)
         except:
             raise RecordFindError(home_id)
+        
+        if home_id and len(home_list) < 1:
+            raise RecordNotFound(home_id)
         
         return home_list
 
     def add_home(self, record):
-        home = Home(record['name'])
+        try:
+            name = record['name']
+        except:
+            raise LackRequestData()
+
+        home = Home(name)
+
         try:
             self._database.add_data(home)
         except:
@@ -31,6 +38,8 @@ class HomeModel():
     def delete_home(self, home_id):
         try:
             home_list = self.get_home(home_id)
+            if len(home_list) < 1:
+                raise RecordNotFound(home_id)
             home = home_list[0]
             self._database.remove_data(home)
         except:
